Code Cleanup - Moved settings forms to subdirectory
