+++ src/containers/Settings/forms/AddWidgetForm.js
@@ -1,8 +1,8 @@
 import React from 'react';
-import classes from './Settings.module.css';
-import FormTemplate from './FormTemplate.js'
-import Option from './Option';
-import * as widgetForms from '../../widgetForms';
+import classes from './Form.module.css';
+import FormTemplate from '../FormTemplate.js'
+import Option from '../Option';
+import * as widgetForms from '../../../widgetForms';
 
 const AddWidgetForm = () => {
     const formList = Object.values(widgetForms);
@@ -10,7 +10,7 @@ const AddWidgetForm = () => {
         <FormTemplate
             title="Add Widget"
             subtitle="Choose a Widget:"
-            formClasses={classes.Settings}
+            formClasses={classes.Form}
         >
             {formList.map(form => <Option key={form.key} opt={form} />)}
         </FormTemplate>
