+++ src/containers/Settings/forms/Form.module.css
@@ -1,0 +1,46 @@
+.Form {
+    display: flex;
+    flex-direction: column;
+    height: 100vh;
+}
+
+.Form header {
+    height: max(140px, 18vh);
+}
+
+.Form section {
+    height: max(620px, 80vh);
+    display: grid;
+    grid-template-columns: repeat(2, 1fr);
+    overflow-y: scroll;
+    -ms-overflow-style: none;  /* Internet Explorer 10+ */
+    scrollbar-width: none;  /* Firefox */
+}
+.Form section::-webkit-scrollbar { 
+    display: none;  /* Safari and Chrome */
+}
+
+.removeRow {
+    display: grid;
+    grid-template-columns: 1fr 2fr 1fr;
+    background-color: rgba(132, 132, 132, 0.32);
+    backdrop-filter: blur(4.8px);
+    text-align: center;
+    border-radius: 12px;
+    margin: 12px auto;
+}
+.removeRow:not(.rowQ) {
+    justify-content: center;
+    align-items: center;
+}
+.rowQ {
+    display: flex;
+    flex-direction: column;
+}
+.rowQ div {
+    display: grid;
+    grid-template-columns: 1fr 1.5fr;
+}
+.rowQ div p {
+    margin: 6px 0;
+}
